Moving snippets into separate files/updating tests
